% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function 2.R
\name{basic}
\alias{basic}
\title{Function 1}
\usage{
basic(res, pred)
}
\arguments{
\item{res}{res should be the response variable of interest}

\item{pred}{pred should be the predictor variable of interest}
}
\description{
Calculates the basic analysis involving simple linear model, plot, and confidence interval
}
\examples{
basic()
}
\keyword{basic}
