% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function 3.R
\name{center}
\alias{center}
\title{Function 3}
\usage{
center(z)
}
\arguments{
\item{z}{should be a conitnuous variable of interest}
}
\description{
Looks closely at the median and mean and will then compare the two.It will determine whether the mean or median is a good use of center.
}
\examples{
center()
#Function3
}
\keyword{mean}
\keyword{median,}
