% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function 1.R
\name{partialR2}
\alias{partialR2}
\title{Function 1}
\usage{
partialR2(r1, r2)
}
\arguments{
\item{r1}{r1 should be the r-squared value of the model that has the predictor variable of interest,}

\item{r2}{r2 should be the r-squared value of the model that does not have the predictor variable of interest}
}
\description{
Calculates the partial R2 value
}
\examples{
partialR2()
}
\keyword{Partial,}
\keyword{R-Squared}
